ranger: rerange file and rename some method.
